+++ src/DownloadCommand.cc
@@ -199,7 +199,8 @@ bool DownloadCommand::executeInternal() {
 	_requestGroup->getDownloadContext()->getPieceHash(segment->getIndex());
       if(_pieceHashValidationEnabled && !expectedPieceHash.empty()) {
 	if(segment->isHashCalculated()) {
-	  logger->debug("Hash is available! index=%zu", segment->getIndex());
+	  logger->debug("Hash is available! index=%lu",
+			static_cast<unsigned long>(segment->getIndex()));
 	  validatePieceHash(segment, expectedPieceHash, segment->getHashString());
 	} else {
 	  _messageDigestContext->digestReset();
