+++ src/SimpleBtMessage.cc
@@ -57,7 +57,8 @@ void SimpleBtMessage::send() {
     size_t msgLength = getMessageLength();
     logger->info(MSG_SEND_PEER_MESSAGE,
 		 cuid, peer->ipaddr.c_str(), peer->port, toString().c_str());
-    logger->debug("msglength = %zu bytes", msgLength);
+    logger->debug("msglength = %lu bytes",
+		  static_cast<unsigned long>(msgLength));
     peerConnection->sendMessage(msg, msgLength);
   } else {
     peerConnection->sendPendingData();
